Use node-frankenstein-wasm directly instead of v8
