+++ js-prototype/constants.js
@@ -1,4 +1,4 @@
-const MAX_BAL_BYTES = 16 // Max balance size in bytes
-const ADD_SIZE_BYTES = 20 // Address size in bytes
-
-
+module.exports = {
+  MAX_BAL_BYTES: 16, // Max balance size in bytes
+  ADD_SIZE_BYTES: 20 // Address size in bytes
+}
