+++ js-prototype/test_runner.js
@@ -5,7 +5,7 @@ const exec = require('child_process').exec
 
 // get the test names
 const tests = fs.readdirSync('../tests').filter((file) => file.split('.')[1] === 'wast')
-const v8_path = (process.env.V8_PATH ? process.evn.V8_PATH : './v8/out/native/') + 'd8'
+const node_path = (process.env.NODEWASM_PATH ? process.env.NODEWASM_PATH : './') + 'node'
 
 // run the tests
 for (let testName of tests) {
@@ -15,7 +15,7 @@ for (let testName of tests) {
     const cmpCmd = `./sexpr-wasm-prototype/out/sexpr-wasm ../tests/${testName}.wast -o ../tests/${testName}.wasm`
 
     // Run Command
-    const runCmd = `${v8_path} -e "const testName = '${testName}'" --expose-wasm v8_test_load.js`
+    const runCmd = `${node_path} -e "const testName = '${testName}'" --expose-wasm node_test_load.js`
 
     exec(cmpCmd + ' && ' + runCmd, (error, stdout, stderr) => {
       t.equals(stdout, 'done\n')
